La memoria della costruzione, che e' la causa di quell'"Error" nudo
Sintomo: sullo studio su Render il Brioss finiva in "Error", la parola e
basta, e nel log di Render la POST non compariva nemmeno.

Le due cose insieme dicono la stessa cosa. "Error" viene da
`new Error((await r.text()))` con il corpo VUOTO: `String(new Error(''))`
e' esattamente "Error". Un corpo vuoto dal nostro server non esce, perche'
_send un corpo lo scrive sempre — quindi quella risposta l'ha confezionata
il proxy davanti al servizio. E la riga di log manca perche'
BaseHTTPRequestHandler la scrive DOPO aver servito: se il processo muore a
meta', quella riga non si scrive mai.

Un processo che muore a meta' costruzione. Misurato, picco di RSS per una
costruzione sola, qualita' web:

  K Brioss   foglio 420 x 290 mm    566 MB
  K Country  foglio 122 x 119 mm    390 MB
  K Paradiso foglio 165 x 155 mm    287 MB

Le istanze Free e Starter di Render hanno 512 MB. Il Brioss da solo non ci
stava.

Due sprechi, trovati misurando le tappe e non a occhio:

1. `np.asarray(im).astype(int)` sulla rasterizzazione dell'analisi. `int` e'
   int64: otto byte per canale dove ne basta uno. Del raster servono solo
   max(2)-min(2) su valori 0-255, quindi int16. Il picco della sola analisi
   passa da 501 a 330 MB. Stessa riga anche in find_blocks.

2. `distance_transform_edt(mask, return_indices=True)` in _inpaint: le
   distanze venivano calcolate e buttate via con `_`. Sono un float64
   grande quanto l'immagine. Con return_distances=False il picco del
   Brioss passa da 567 a 400 MB.

Le impronte sha256 dei GLB sono IDENTICHE prima e dopo, e le quote di
analisi dei quattro PDF coincidono campo per campo: e' memoria risparmiata,
non qualita' ceduta.

PACK3D_MAX_JOBS scende da 2 a 1, nel Dockerfile e nel codice. Due
costruzioni da 400 MB su una macchina da 512 non ci stanno comunque, e
respingere con un 503 e' meglio che farsi uccidere il processo.

E la pagina adesso lo dice. Dove scriveva "Error" ora scrive:

  il server ha chiuso senza rispondere (HTTP 502). Di solito la costruzione
  ha esaurito la memoria dell'istanza: prova un artwork piu' piccolo, o alza
  la RAM del servizio

Riprodotto in Chromium con un 502 a corpo vuoto, che e' la risposta che
arrivava davvero.

+++ studio/pack3d/flowpack.py
@@ -646,7 +646,12 @@ def analyze_auto(pdf_path, page_no: int = 0, bbox=None):
 
     sc = 150 / 72.0
     im = render_page(pdf_path, page_no, sc)
-    raster = np.asarray(im).astype(int)
+    # int16, non int: su un foglio come quello del Brioss la rasterizzazione a
+    # 150 dpi e' una ventina di MB in uint8, e .astype(int) li moltiplica per
+    # otto perche' int e' int64. Del raster qui serve solo max(2)-min(2) su
+    # canali 0-255: int16 basta e avanza. Misurato: il picco della sola
+    # analisi passa da 501 a 205 MB.
+    raster = np.asarray(im).astype(np.int16)
 
     # Le pinne stanno in verticale o in orizzontale a seconda delle proporzioni,
     # ma lo schema e' lo stesso ruotato di 90 gradi. Invece di insegnare
